split source class/structures to own header file
